Extract type casting template functions into a type cast Jinja2 extension
