+++ homeassistant/helpers/template/extensions/__init__.py
@@ -12,6 +12,7 @@ from .labels import LabelExtension
 from .math import MathExtension
 from .regex import RegexExtension
 from .string import StringExtension
+from .type_cast import TypeCastExtension
 
 __all__ = [
     "AreaExtension",
@@ -26,4 +27,5 @@ __all__ = [
     "MathExtension",
     "RegexExtension",
     "StringExtension",
+    "TypeCastExtension",
 ]
